Add files via upload
This project focuses on understanding customer behavior in an e-commerce setting using a dataset containing user and interaction data. The primary goal is to determine which factors most influence a customer's yearly amount spent and whether the company should invest more in improving its website or mobile app experience to boost revenue.